fixed image update  bug

+++ pomodoro.py
@@ -8,20 +8,7 @@ from datetime import datetime
 
 PATH = os.path.dirname(os.path.abspath(__file__))
 
-# Global Inky Properties
-inky_display = InkyPHAT("red")
-inky_display.rotation = 0 # avoid unnecessary swap
-
-# set pomodoro cycle length
-pomodoro = 25 * 60 # typical length
-small_break = 5 * 60
-long_break = 25 * 60
-cycle_length = (pomodoro + small_break) * 3 +  pomodoro + long_break
-
-# set fonts
-font = ImageFont.truetype(FredokaOne, 32)
-
-def format_line(msg):
+def format_line(font, msg):
     lines = []
     w, h = font.getsize(msg)
     if w <= inky_display.WIDTH:
@@ -40,12 +27,16 @@ def format_line(msg):
     return lines
 
 def get_pomodoro_time(start_time):
+    # set pomodoro cycle length
+    pomodoro = 25 * 60 # typical length
+    small_break = 5 * 60
+    long_break = 25 * 60
+    cycle_length = (pomodoro + small_break) * 3 +  pomodoro + long_break
     curr_time = int(datetime.utcnow().timestamp()) % 86400
     time_since_start = curr_time - start_time
     completed_cycles = int(time_since_start // cycle_length)
     pt_in_cycle = time_since_start - completed_cycles * cycle_length
     num_tomato = int(pt_in_cycle // (pomodoro + small_break))
-
     if int(pt_in_cycle % (pomodoro + small_break)  >= pomodoro) and (num_tomato < 3):
         print(num_tomato, "break time")
         return (num_tomato, "break time!")
@@ -56,7 +47,7 @@ def get_pomodoro_time(start_time):
         print(num_tomato, "still working")
         return (num_tomato, "still working")
 
-def get_tomato_image(image_num):
+def get_tomato_image(inky_display, image_num):
     """
     Use PIL library to open tomato image and transpose for inky display
     """
@@ -67,8 +58,9 @@ def get_tomato_image(image_num):
     canvas = canvas.transpose(Image.ROTATE_90)
     return canvas
 
-def get_text_image(break_text):
-    lines = format_line(break_text)
+def get_text_image(inky_display, break_text):
+    font = ImageFont.truetype(FredokaOne, 32)
+    lines = format_line(font, break_text)
     _, line_height = font.getsize(lines[0])
     centered_y = (inky_display.HEIGHT / 2) - ((line_height * len(lines)) / 2)
     height_counter = centered_y
@@ -84,37 +76,41 @@ def get_text_image(break_text):
     canvas = img.transpose(Image.ROTATE_90)
     return canvas
 
-def check_display(tomato, cycle, start_time):
+def check_display(inky_display, tomato, cycle, start_time):
     num_tomato, text = get_pomodoro_time(start_time)
     # check if status has changed
-    if num_tomato == tomato and text == cycle:
+    # if num_tomato == tomato and text == cycle:
+    if False:
         return
     else:
         if text == "still working":
-            img = get_tomato_image(num_tomato)
+            img = get_tomato_image(inky_display, num_tomato)
         else:
-            img = get_text_image(text)
+            img = get_text_image(inky_display, text)
         # update status.json
-        out_dict = {'num_tomato': num_tomato, 'status_cycle': text}
-        with open('status.json', 'w') as fh:
+        out_dict = {'num_tomato': num_tomato, 'status_cycle': text, 'start_time': start_time}
+        with open(PATH + '/status.json', 'w') as fh:
             json.dump(out_dict, fh)
-        inky_display.set_img(img)
+        inky_display.set_image(img)
         inky_display.show()
 
+# Global Inky Properties
+inky_display = InkyPHAT("red")
+inky_display.rotation = 0 # avoid unnecessary swap
 
-try:
-    with open('status.json', 'r') as fh:
+if os.path.exists(PATH + '/status.json'):
+    with open(PATH + '/status.json', 'r') as fh:
         status = json.load(fh)
         status_tomato = status["num_tomato"]
-        status_cycle = status["pt_in_cycle"]
+        status_cycle = status["status_cycle"]
         start_time = status['start_time']
-except:
-    with open('status.json', 'w') as fh:
-        status_tomato = 0
-        status_cycle = "still working"
-        start_time = int(datetime.utcnow().timestamp()) % 86400
-        status = {'num_tomato': 0, 'status_cycle': 'still working', 'start_time': start_time}
+        check_display(inky_display, status_tomato, status_cycle, start_time)
+else:
+    status_tomato = 0
+    status_cycle = "still working"
+    start_time = int(datetime.utcnow().timestamp()) % 86400
+    check_display(inky_display, status_tomato, status_cycle, start_time)
+    status = {'num_tomato': 0, 'status_cycle': 'still working', 'start_time': start_time}
+    with open(PATH + '/status.json', 'w') as fh:
         json.dump(status, fh)
-
-check_display(status_tomato, status_cycle, start_time)
